fix warning

+++ cbase/arena.c
@@ -382,7 +382,7 @@ main(void) {
         int64 total_pushed = 0;
 
         for (int32 i = 0; i < LENGTH(objs); i += 1) {
-            int64 size = ALIGN(1 + (int64)(rand_int() % 10000u));
+            int64 size = ALIGN(1 + (rand_int() % 10000u));
             ASSERT((objs[i] = arena_push(arena, size)));
 
             total_size += size;
